Move `FilteredNGramStat` to its own module
